ModulDynamicTestingDSAPIBackend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
httpx==0.27.0
pydantic==2.6.3
pydantic-settings==2.2.1
pandas==2.2.1
openpyxl==3.1.2
tqdm==4.66.2
torch>=2.0.0
transformers>=4.38.0
accelerate>=0.27.0
langchain>=0.1.12
langchain-ollama
langchain-core
natasha>=1.6.0
pymorphy3==1.2.1
pymorphy3-dicts-ru
pyspellchecker==0.8.1
langdetect==1.0.9
setuptools
Pillow>=10.2.0
beautifulsoup4==4.12.3